refactor: merge advanced variants into base modules, consolidate layouts and performance
