+++ .superpowers/sdd/task-7-report.md
@@ -1,0 +1,85 @@
+# Task 7 Report: Split monolithic data module into src/data/
+
+**Status:** ISSUES FOUND — see resolution required
+
+## Spec Compliance
+
+| Requirement | Status | Detail |
+|---|---|---|
+| 4 sub-modules created (validation, transforms, advanced, pipeline) | PASS | 4 files in `src/data/` |
+| Old monolithic files deleted (`mintwaterfall-data.ts`, `mintwaterfall-advanced-data.ts`) | PASS | Both deleted |
+| `createDataProcessor()` returns same API surface | PASS | All methods present, signatures match |
+| Standalone helpers exported | PASS | `createRevenueWaterfall`, `createTemporalWaterfall`, `createVarianceWaterfall`, `groupWaterfallData`, `createComparisonWaterfall`, `transformTransactionData` |
+| `financialReducers` preserved | PASS | Identical object in `pipeline.ts:302` |
+| `d3DataUtils` preserved | PASS | Identical object in `pipeline.ts:322` |
+| Import paths use `.js` extensions for ESM | PASS | All inter-module imports use `.js` |
+| Code copied verbatim, only import paths changed | PASS | Functions match original except import changes |
+| `tsc --noEmit` passes | PASS | Clean exit 0 |
+| No dependency bumps | PASS | Only file changes |
+| All 4 build formats must build | UNVERIFIED | `tsc --noEmit` passes; full build not tested |
+| All original exports preserved | FAIL | See Issues #1 and #2 |
+
+## Strengths
+
+1. **Clean module boundaries.** Validation (types + `validateData` + `getDataSummary`), transforms (12 data manipulation functions), advanced (9 D3 operations), and pipeline (composition + standalone helpers) are logically separated and independently importable.
+
+2. **Verbatim code extraction.** All function bodies were copied exactly from the original `src/mintwaterfall-data.ts`. No logic changes or refactoring drift.
+
+3. **ESM-transition-ready.** All cross-module imports use `.js` extensions (`"./validation.js"`, `"./transforms.js"`, `"./advanced.js"`), ready for Node's native ESM.
+
+4. **`createDataProcessor()` API surface preserved.** The factory function returns a `DataProcessor` object with all original methods, and `dataProcessor` (default instance) is exported for backward compatibility.
+
+5. **Type exports proxied.** Core types (`StackItem`, `DataItem`, etc.) are re-exported from `pipeline.ts` so consumers importing from a single entry point continue to work.
+
+## Issues
+
+### Issue #1 (HIGH): Three type exports dropped from pipeline.ts
+
+**Affected:** `TimeScaleOptions`, `OrdinalScaleOptions`, `BandScaleOptions`
+
+These were top-level `export interface` declarations in the original `src/mintwaterfall-data.ts` (lines ~41-55). They are exported from `src/data/validation.ts` but are NOT re-exported from `src/data/pipeline.ts`. Any consumer that previously imported these types from the monolithic data module will get a TypeScript error.
+
+**Fix:** Add to the pipeline.ts re-export block (line 47-60):
+```typescript
+export {
+    ...
+    TimeScaleOptions,
+    OrdinalScaleOptions,
+    BandScaleOptions,
+    ...
+};
+```
+
+### Issue #2 (HIGH): `mintwaterfall-advanced-data.ts` content deleted without migration
+
+**Affected exports:** `createAdvancedDataProcessor`, `createWaterfallSequenceAnalyzer`, `createWaterfallTickGenerator`, `SequenceAnalysis`, `DataMergeOptions`, `TickGenerationOptions`, `DataOrderingOptions`
+
+The original `src/mintwaterfall-advanced-data.ts` (1034 lines) was deleted, but its content was NOT extracted into any of the new modules. The task brief instructed extracting from `src/mintwaterfall-data.ts` lines 541-843 for `advanced.ts`, which covers the D3 data operations — not the sequence/tick/ordering utilities from the separate `-advanced-data.ts` file.
+
+**Impact:** `src/index.js:65-69` still imports these three functions from the now-deleted file. Runtime will throw `ERR_MODULE_NOT_FOUND`. This violates the "No public API breakage" global constraint.
+
+**Resolution:** Either:
+- (a) Create `src/data/sequence.ts` with the migrated content from `mintwaterfall-advanced-data.ts`, or
+- (b) Explicitly declare these exports as intentionally removed (and update index.js accordingly) if they're being superseded by the new pipeline architecture.
+
+### Issue #3 (LOW): Unused import in validation.ts
+
+`src/data/validation.ts:2` imports `import * as d3 from "d3"` but neither `validateData()` nor `getDataSummary()` uses `d3`. The import was copied from the brief's template but is dead code.
+
+### Issue #4 (LOW): `loadData` and `transformToWaterfallFormat` no longer standalone
+
+In the original `mintwaterfall-data.ts`, these were `export function` at top level (lines 125, 184). In the new structure, they are exported from `transforms.ts` but NOT re-exported as standalone from `pipeline.ts`. They are only accessible through `dataProcessor.loadData()` / `dataProcessor.transformToWaterfallFormat()`.
+
+Since `index.js` does not import them standalone (only through `createDataProcessor`), this has no practical impact. However, any third-party consumer directly importing these functions from the data module would break. Low risk, but note it.
+
+## Assessment
+
+**CONDITIONAL PASS** — 2 HIGH issues must be resolved.
+
+The core refactoring (splitting the 1100-line `mintwaterfall-data.ts` into 4 focused sub-modules) was executed faithfully to the brief. TypeScript compiles cleanly, function signatures match the original, and the `createDataProcessor()` API surface is preserved.
+
+However, two API gaps exist:
+1. Three scale-related type exports (`TimeScaleOptions`, `OrdinalScaleOptions`, `BandScaleOptions`) missing from pipeline.ts re-exports.
+2. The entire `mintwaterfall-advanced-data.ts` module was deleted without content migration, removing `createAdvancedDataProcessor`, `createWaterfallSequenceAnalyzer`, and `createWaterfallTickGenerator` from the public API.
+
+Issue #1 is a one-line fix (add to re-export block). Issue #2 requires a decision: migrate the content or formally deprecate the exports in index.js.
