--- src/types/js-modules.d.ts
@@ -1,25 +1,0 @@
-// Type declarations for JavaScript modules during gradual TypeScript migration
-
-declare module "../mintwaterfall-brush.js" {
-    export function createBrushSystem(): any;
-}
-
-declare module "../mintwaterfall-accessibility.js" {
-    export function createAccessibilitySystem(): any;
-}
-
-declare module "../mintwaterfall-tooltip.js" {
-    export function createTooltipSystem(): any;
-}
-
-declare module "../mintwaterfall-export.js" {
-    export function createExportSystem(): any;
-}
-
-declare module "../mintwaterfall-zoom.js" {
-    export function createZoomSystem(): any;
-}
-
-declare module "../mintwaterfall-performance.js" {
-    export function createPerformanceManager(): any;
-}
